refactor: float -> double

+++ src/main/java/app/Index.java
@@ -11,7 +11,7 @@ public class Index {
 
     try {
       System.out.println("Suma de matrices");
-      float[][] sum1 = { { 2, -3, 5 }, { 4, 1, -7 } }, sum2 = { { 1, 0, 2 }, { -3, 5, 8 } };
+      double[][] sum1 = { { 2, -3, 5 }, { 4, 1, -7 } }, sum2 = { { 1, 0, 2 }, { -3, 5, 8 } };
       Matrix matrixSum1 = new Matrix(sum1), matrixSum2 = new Matrix(sum2);
       System.out.println(Arrays.deepToString(controller.sum(matrixSum1, matrixSum2).getData())); 
     } catch (Exception e) {
@@ -21,7 +21,7 @@ public class Index {
 
     try {
       System.out.println("Multiplicación por escalar");
-      float[][] scalarMultiplication = { { 1, -2, 3 }, { 0, 1, 8 } };
+      double[][] scalarMultiplication = { { 1, -2, 3 }, { 0, 1, 8 } };
       Matrix matrixScalarMult = new Matrix(scalarMultiplication);
       System.out.println(Arrays.deepToString(controller.multiplicationByScalar(matrixScalarMult, -5).getData()));  
     } catch (Exception e) {
@@ -31,7 +31,7 @@ public class Index {
 
     try {
       System.out.println("Multiplicación de matrices");
-      float[][] mult1 = { { 1, 0, 2 }, { -1, 3, 1 } }, mult2 = { { 3, 1 }, { 2, 1 }, { 1, 0 } };
+      double[][] mult1 = { { 1, 0, 2 }, { -1, 3, 1 } }, mult2 = { { 3, 1 }, { 2, 1 }, { 1, 0 } };
       Matrix matrixMult1 = new Matrix(mult1), matrixMult2 = new Matrix(mult2);
       System.out.println(Arrays.deepToString(controller.multiplicationOfTwoMatrices(matrixMult1, matrixMult2).getData())); 
     } catch (Exception e) {
@@ -41,7 +41,7 @@ public class Index {
 
     try {
       System.out.println("Matriz inversa por Gauss-Jordan");
-      float[][] gaussJordan = { { 1, -6, 2 },  { 2, -2, -1 }, { 1, -3, -5 } };
+      double[][] gaussJordan = { { 1, -6, 2 },  { 2, -2, -1 }, { 1, -3, -5 } };
       Matrix matrix = new Matrix(gaussJordan);
       System.out.println(Arrays.deepToString(controller.getInverseByGaussJordan(matrix).getData())); 
     } catch (Exception e) {
@@ -50,12 +50,12 @@ public class Index {
 
     try {
       System.out.println("Sistema de ecuaciones por Gauss-Jordan");
-      float[][] gaussJordan = {
-        { 3, (float)-0.1, (float)-0.2 },
-        { (float)0.1, 7, (float)-0.3 },
-        { (float)0.3, (float)-0.2, 10 }
+      double[][] gaussJordan = {
+        { 3, -0.1, -0.2 },
+        { 0.1, 7, -0.3 },
+        { 0.3, -0.2, 10 }
       };
-      float[] independentValues = { (float)7.85, (float)-19.3, (float)71.4 };
+      double[] independentValues = { 7.85, -19.3, 71.4 };
       Matrix matrix = new Matrix(gaussJordan);
       System.out.println(Arrays.toString(controller.getSystemOfEquations(matrix, independentValues)));
     } catch (Exception e) {
